# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface PricingTier {
  name: string;
  price: string;
  rawPrice: number;
  desc: string;
  popular?: boolean;
  features: string[];
}

export interface SiteSettings {
  whatsappNumber: string;
  paypalUsername: string;
  currency: string;
  consultationEmail: string;
  businessName: string;
  /**
   * @minItems 3
   * @maxItems 3
   */
  tiers: PricingTier[];
}

export interface NewsletterSubscribeBody {
  email: string;
  /** @maxLength 120 */
  name?: string;
}

export interface NewsletterSubscriber {
  id: number;
  email: string;
  name?: string | null;
  createdAt: string;
}

export interface CreateBookingBody {
  reference: string;
  planName: string;
  planPrice: string;
  /**
   * @minimum 0
   * @maximum 2
   */
  planIndex: number;
  clientName: string;
  clientEmail: string;
  clientWhatsapp: string;
  /** @maxLength 2000 */
  concern: string;
  appointmentDate?: string | null;
  appointmentTime?: string | null;
}

export interface Booking {
  id: number;
  reference: string;
  planName: string;
  planPrice: string;
  planIndex: number;
  clientName: string;
  clientEmail: string;
  clientWhatsapp: string;
  concern: string;
  appointmentDate?: string | null;
  appointmentTime?: string | null;
  status: string;
  createdAt: string;
}