# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Pocketpill API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: settings
    description: Site settings (admin only)
  - name: newsletter
    description: Newsletter subscriptions
  - name: bookings
    description: Consultation bookings

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriber"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /newsletter/subscribers:
    get:
      operationId: getNewsletterSubscribers
      tags: [newsletter]
      summary: List all newsletter subscribers (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsletterSubscriber"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Reference already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List all bookings (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    PricingTier:
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        rawPrice:
          type: number
        desc:
          type: string
        popular:
          type: boolean
        features:
          type: array
          items:
            type: string
      required: [name, price, rawPrice, desc, features]

    SiteSettings:
      type: object
      properties:
        whatsappNumber:
          type: string
        paypalUsername:
          type: string
        currency:
          type: string
        consultationEmail:
          type: string
        businessName:
          type: string
        tiers:
          type: array
          items:
            $ref: "#/components/schemas/PricingTier"
          minItems: 3
          maxItems: 3
      required:
        [whatsappNumber, paypalUsername, currency, consultationEmail, businessName, tiers]

    NewsletterSubscribeBody:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 120
      required: [email]

    NewsletterSubscriber:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, createdAt]

    CreateBookingBody:
      type: object
      properties:
        reference:
          type: string
        planName:
          type: string
        planPrice:
          type: string
        planIndex:
          type: integer
          minimum: 0
          maximum: 2
        clientName:
          type: string
        clientEmail:
          type: string
          format: email
        clientWhatsapp:
          type: string
        concern:
          type: string
          maxLength: 2000
        appointmentDate:
          type: string
          nullable: true
        appointmentTime:
          type: string
          nullable: true
      required:
        [reference, planName, planPrice, planIndex, clientName, clientEmail, clientWhatsapp, concern]

    Booking:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        planName:
          type: string
        planPrice:
          type: string
        planIndex:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
        clientWhatsapp:
          type: string
        concern:
          type: string
        appointmentDate:
          type: string
          nullable: true
        appointmentTime:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        [id, reference, planName, planPrice, planIndex, clientName, clientEmail, clientWhatsapp, concern, status, createdAt]