# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get site settings
 */
export const getSettingsResponseTiersMin = 3;
export const getSettingsResponseTiersMax = 3;

export const GetSettingsResponse = zod.object({
  whatsappNumber: zod.string(),
  paypalUsername: zod.string(),
  currency: zod.string(),
  consultationEmail: zod.string(),
  businessName: zod.string(),
  tiers: zod
    .array(
      zod.object({
        name: zod.string(),
        price: zod.string(),
        rawPrice: zod.number(),
        desc: zod.string(),
        popular: zod.boolean().optional(),
        features: zod.array(zod.string()),
      }),
    )
    .min(getSettingsResponseTiersMin)
    .max(getSettingsResponseTiersMax),
});

/**
 * @summary Update site settings
 */
export const updateSettingsBodyTiersMin = 3;
export const updateSettingsBodyTiersMax = 3;

export const UpdateSettingsBody = zod.object({
  whatsappNumber: zod.string(),
  paypalUsername: zod.string(),
  currency: zod.string(),
  consultationEmail: zod.string(),
  businessName: zod.string(),
  tiers: zod
    .array(
      zod.object({
        name: zod.string(),
        price: zod.string(),
        rawPrice: zod.number(),
        desc: zod.string(),
        popular: zod.boolean().optional(),
        features: zod.array(zod.string()),
      }),
    )
    .min(updateSettingsBodyTiersMin)
    .max(updateSettingsBodyTiersMax),
});

export const updateSettingsResponseTiersMin = 3;
export const updateSettingsResponseTiersMax = 3;

export const UpdateSettingsResponse = zod.object({
  whatsappNumber: zod.string(),
  paypalUsername: zod.string(),
  currency: zod.string(),
  consultationEmail: zod.string(),
  businessName: zod.string(),
  tiers: zod
    .array(
      zod.object({
        name: zod.string(),
        price: zod.string(),
        rawPrice: zod.number(),
        desc: zod.string(),
        popular: zod.boolean().optional(),
        features: zod.array(zod.string()),
      }),
    )
    .min(updateSettingsResponseTiersMin)
    .max(updateSettingsResponseTiersMax),
});

/**
 * @summary Subscribe to newsletter
 */
export const subscribeNewsletterBodyNameMax = 120;

export const SubscribeNewsletterBody = zod.object({
  email: zod.string().email(),
  name: zod.string().max(subscribeNewsletterBodyNameMax).optional(),
});

/**
 * @summary List all newsletter subscribers (admin)
 */
export const GetNewsletterSubscribersResponseItem = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetNewsletterSubscribersResponse = zod.array(
  GetNewsletterSubscribersResponseItem,
);

/**
 * @summary Create a booking
 */
export const createBookingBodyPlanIndexMin = 0;
export const createBookingBodyPlanIndexMax = 2;

export const createBookingBodyConcernMax = 2000;

export const CreateBookingBody = zod.object({
  reference: zod.string(),
  planName: zod.string(),
  planPrice: zod.string(),
  planIndex: zod
    .number()
    .min(createBookingBodyPlanIndexMin)
    .max(createBookingBodyPlanIndexMax),
  clientName: zod.string(),
  clientEmail: zod.string().email(),
  clientWhatsapp: zod.string(),
  concern: zod.string().max(createBookingBodyConcernMax),
  appointmentDate: zod.string().nullish(),
  appointmentTime: zod.string().nullish(),
});

/**
 * @summary List all bookings (admin)
 */
export const GetBookingsResponseItem = zod.object({
  id: zod.number(),
  reference: zod.string(),
  planName: zod.string(),
  planPrice: zod.string(),
  planIndex: zod.number(),
  clientName: zod.string(),
  clientEmail: zod.string(),
  clientWhatsapp: zod.string(),
  concern: zod.string(),
  appointmentDate: zod.string().nullish(),
  appointmentTime: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetBookingsResponse = zod.array(GetBookingsResponseItem);