# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Booking {
  id: number;
  reference: string;
  planName: string;
  planPrice: string;
  planIndex: number;
  clientName: string;
  clientEmail: string;
  clientWhatsapp: string;
  concern: string;
  appointmentDate?: string | null;
  appointmentTime?: string | null;
  status: string;
  createdAt: Date;
}