# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateBookingBody {
  reference: string;
  planName: string;
  planPrice: string;
  /**
   * @minimum 0
   * @maximum 2
   */
  planIndex: number;
  clientName: string;
  clientEmail: string;
  clientWhatsapp: string;
  /** @maxLength 2000 */
  concern: string;
  appointmentDate?: string | null;
  appointmentTime?: string | null;
}