# lib/api-zod/src/generated/types/newsletterSubscribeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */

export interface NewsletterSubscribeBody {
  email: string;
  /** @maxLength 120 */
  name?: string;
}