# lib/api-zod/src/generated/types/newsletterSubscriber.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */

export interface NewsletterSubscriber {
  id: number;
  email: string;
  name?: string | null;
  createdAt: Date;
}