# lib/api-zod/src/generated/types/pricingTier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */

export interface PricingTier {
  name: string;
  price: string;
  rawPrice: number;
  desc: string;
  popular?: boolean;
  features: string[];
}