# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pocketpill API specification
 * OpenAPI spec version: 0.2.0
 */
import type { PricingTier } from "./pricingTier";

export interface SiteSettings {
  whatsappNumber: string;
  paypalUsername: string;
  currency: string;
  consultationEmail: string;
  businessName: string;
  /**
   * @minItems 3
   * @maxItems 3
   */
  tiers: PricingTier[];
}